Created Base Class. Added more itens to .gitignore. Fixed to only add credentials when have keys.
