+++ .gitignore
@@ -1,1 +1,6 @@
-vendor/
+/composer.lock
+/node_modules
+/vendor
+.env
+/.idea
+/.vscode
